chore(rules): factor BuildKit syntax pragma helper into shared file
PRs #682 (prefer-gemfile-bind-mounts) and #683 (prefer-network-none-install)
both introduced identical `dockerfileSyntaxBuildKitMarkers` and
`hasBuildKitSyntaxPragma` helpers. Merging them in parallel produced a
duplicate-symbol build break on `main`.

Move the helper to `buildkit_pragma.go` and remove the duplicates from
both rules. No behavior change.
