+++ internal/rules/tally/ruby/prefer_network_none_install.go
@@ -1,8 +1,6 @@
 package ruby
 
 import (
-	"strings"
-
 	"github.com/moby/buildkit/frontend/dockerfile/instructions"
 
 	"github.com/wharflab/tally/internal/facts"
@@ -11,29 +9,6 @@ import (
 	"github.com/wharflab/tally/internal/semantic"
 	"github.com/wharflab/tally/internal/stagename"
 )
-
-// dockerfileSyntaxBuildKitMarkers identifies BuildKit-frontend syntax
-// pragmas. Cache and bind mounts require one of these.
-var dockerfileSyntaxBuildKitMarkers = []string{"docker/dockerfile", "dockerfile/labs"}
-
-// hasBuildKitSyntaxPragma reports whether the Dockerfile carries a
-// `# syntax=docker/dockerfile:1` (or compatible) directive at its top.
-func hasBuildKitSyntaxPragma(input rules.LintInput) bool {
-	for line := range strings.SplitSeq(string(input.Source), "\n") {
-		trimmed := strings.TrimSpace(line)
-		if !strings.HasPrefix(trimmed, "#") {
-			break
-		}
-		if strings.Contains(trimmed, "syntax=") {
-			for _, m := range dockerfileSyntaxBuildKitMarkers {
-				if strings.Contains(trimmed, m) {
-					return true
-				}
-			}
-		}
-	}
-	return false
-}
 
 // PreferNetworkNoneInstallRuleCode is the full rule code.
 const PreferNetworkNoneInstallRuleCode = rules.TallyRulePrefix + "ruby/prefer-network-none-install"
